refactor: statistics 관련 파일을 feature/statistics 폴더 내로 이동
