+++ lib/features/statistics/domain/repositories/statistics_repository.dart
@@ -1,4 +1,5 @@
-import 'package:mongbi_app/domain/entities/statistics.dart';
+
+import 'package:mongbi_app/features/statistics/domain/entities/statistics.dart';
 
 abstract interface class StatisticsRepository {
   Future<Statistics?> fetchMonthStatistics(DateTime dateTime);
