+++ lib/features/statistics/presentation/models/statistics_model.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/domain/entities/statistics.dart';
+import 'package:mongbi_app/features/statistics/domain/entities/statistics.dart';
 
 class StatisticsModel {
   StatisticsModel({this.month, this.year});
