+++ lib/features/statistics/providers/statistics_provider.dart
@@ -1,16 +1,16 @@
 import 'dart:io';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/data/data_sources/remote_statistics_data_source.dart';
-import 'package:mongbi_app/data/data_sources/statistics_data_source.dart';
-import 'package:mongbi_app/data/repositories/remote_statistics_repository.dart';
-import 'package:mongbi_app/domain/repositories/statistics_repository.dart';
-import 'package:mongbi_app/domain/use_cases/fetch_month_statistics_use_case.dart';
-import 'package:mongbi_app/domain/use_cases/fetch_year_statistics_use_case.dart';
-import 'package:mongbi_app/presentation/statistics/models/picker_model.dart';
-import 'package:mongbi_app/presentation/statistics/models/statistics_model.dart';
-import 'package:mongbi_app/presentation/statistics/view_models/picker_view_model.dart';
-import 'package:mongbi_app/presentation/statistics/view_models/statistics_view_model.dart';
+import 'package:mongbi_app/features/statistics/data/data_sources/remote_statistics_data_source.dart';
+import 'package:mongbi_app/features/statistics/data/data_sources/statistics_data_source.dart';
+import 'package:mongbi_app/features/statistics/data/repositories/remote_statistics_repository.dart';
+import 'package:mongbi_app/features/statistics/domain/repositories/statistics_repository.dart';
+import 'package:mongbi_app/features/statistics/domain/use_cases/fetch_month_statistics_use_case.dart';
+import 'package:mongbi_app/features/statistics/domain/use_cases/fetch_year_statistics_use_case.dart';
+import 'package:mongbi_app/features/statistics/presentation/models/picker_model.dart';
+import 'package:mongbi_app/features/statistics/presentation/models/statistics_model.dart';
+import 'package:mongbi_app/features/statistics/presentation/view_models/picker_view_model.dart';
+import 'package:mongbi_app/features/statistics/presentation/view_models/statistics_view_model.dart';
 import 'package:mongbi_app/providers/core_providers.dart';
 
 final _statisticsDataSourceProvider = Provider<StatisticsDataSource>((ref) {
