+++ test/data/statistics_repository_test.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:mongbi_app/data/data_sources/statistics_data_source.dart';
-import 'package:mongbi_app/data/dtos/statistics_dto.dart';
-import 'package:mongbi_app/data/repositories/remote_statistics_repository.dart';
-import 'package:mongbi_app/domain/entities/statistics.dart';
+import 'package:mongbi_app/features/statistics/data/data_sources/statistics_data_source.dart';
+import 'package:mongbi_app/features/statistics/data/dtos/statistics_dto.dart';
+import 'package:mongbi_app/features/statistics/data/repositories/remote_statistics_repository.dart';
+import 'package:mongbi_app/features/statistics/domain/entities/statistics.dart';
 
 class MockRemoteDataSource extends Mock implements StatisticsDataSource {}
